feat: Add animation and configurable NOC notification interval
Introduces `framer-motion` for UI animations on the public dashboard, enhancing user experience. The `NOTIFY_THRESHOLD` configuration is now parsed as a float, allowing for more granular control over the interval for NOC notifications. The background task for NOC notifications will now use this threshold to determine its execution frequency, ensuring it runs at least every 10 minutes.

+++ src/server/config.ts
@@ -35,4 +35,4 @@ try {
 export const DASHBOARD_CONFIGS = dashboards;
 
 export const NOC_EMAIL = process.env.NOC_EMAIL;
-export const NOTIFY_THRESHOLD = parseInt(process.env.NOTIFY_THRESHOLD || '0');
+export const NOTIFY_THRESHOLD = parseFloat(process.env.NOTIFY_THRESHOLD || '0');
